fix: keep tool api routes available locally
